+++ sentry-android-core/src/test/java/io/sentry/android/core/CustomCachedApplyScopeDataHint.kt
@@ -1,0 +1,6 @@
+package io.sentry.android.core
+
+import io.sentry.hints.ApplyScopeData
+import io.sentry.hints.Cached
+
+class CustomCachedApplyScopeDataHint : Cached, ApplyScopeData