+++ sentry/src/main/java/io/sentry/SentryTracer.java
@@ -141,7 +141,9 @@ public final class SentryTracer implements ITransaction {
       this.baggage = new Baggage(hub.getOptions().getLogger());
     }
 
-    if (transactionPerformanceCollector != null) {
+    // We are currently sending the performance data only in profiles, so there's no point in
+    // collecting them if a profile is not sampled
+    if (transactionPerformanceCollector != null && Boolean.TRUE.equals(isProfileSampled())) {
       transactionPerformanceCollector.start(this);
     }
 
@@ -437,7 +439,7 @@ public final class SentryTracer implements ITransaction {
 
     this.finishStatus = FinishStatus.finishing(status);
     if (!root.isFinished() && (!waitForChildren || hasAllChildrenFinished())) {
-      PerformanceCollectionData performanceCollectionData = null;
+      List<PerformanceCollectionData> performanceCollectionData = null;
       if (transactionPerformanceCollector != null) {
         performanceCollectionData = transactionPerformanceCollector.stop(this);
       }
@@ -448,6 +450,9 @@ public final class SentryTracer implements ITransaction {
             hub.getOptions()
                 .getTransactionProfiler()
                 .onTransactionFinish(this, performanceCollectionData);
+      }
+      if (performanceCollectionData != null) {
+        performanceCollectionData.clear();
       }
 
       // try to get the high precision timestamp from the root span