+++ sentry/src/main/java/io/sentry/ShutdownHookIntegration.java
@@ -41,7 +41,17 @@ public final class ShutdownHookIntegration implements Integration, Closeable {
   @Override
   public void close() throws IOException {
     if (thread != null) {
-      runtime.removeShutdownHook(thread);
+      try {
+        runtime.removeShutdownHook(thread);
+      } catch (IllegalStateException e) {
+        @Nullable final String message = e.getMessage();
+        // https://github.com/openjdk/jdk/blob/09b8a1959771213cb982d062f0a913285e4a0c6e/src/java.base/share/classes/java/lang/ApplicationShutdownHooks.java#L83
+        if (message != null && message.equals("Shutdown in progress")) {
+          // ignore
+        } else {
+          throw e;
+        }
+      }
     }
   }